commit2

+++ src/main/java/com/multipledbs/Employee/EmployeeRepository.java
@@ -3,8 +3,6 @@ package com.multipledbs.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
-
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
